The traslation "直譯器" for "interpreter" didn't get merged into 3.2.7 … (#1170)
* The translattion of "Interpreter" for zh_TW should be  "直譯器"

The translation of "interpreter" in current version is "直譯器" or "直釋器", it should be unified to "直譯器". That's what we call "interpreter" here in Taiwan--the place where Tradition Chinese is mainly used.

* The traslation "直譯器" for "interpreter" didn't get merged into 3.2.7 

The traslation "直譯器" for "interpreter"  seems not be merged into 3.2.7.I've edit the terms on POEDITOR, but the po file didn't updated.