Hide extra dependencies in plugin manager
This patch hides the dependencies of a plug-in which refer to
`extra == *`.

These dependencies can be confusing for the user as they are actually
not installed. For example, a plug-in might use `black` to format the
code and mark it as `dev` dependency, but Thonny will not install it
when acquiring the plug-in.

We added `packaging` to requirements instead of importing
`packaging.markers` from `pkg_resources` or `pip` (as library). This is
much easier to trace instead of using `__import__` builtin.