Use UDisks to find volumes on Linux
This should generally work better on Linux than using the mount command directly.
UDisks is the utility that all major Linux desktop distributions use to handle mounting filesystems in userspace.
UDisks is accessible through a D-Bus API on the system bus.

D-Bus is complex and requires adding a dependency on a Python library for handling asynchronous communication with the bus.
I chose dbus-next since it is actively developed, modern, and doesn't require a glib main loop.
Since Thonny is not a Gtk application, it makes more sense to use native Python async functionality instead.

Inside a Flatpak, it's necessary to use UDisks via the D-Bus API to mount and check mounted filesystems.
The Flatpak sandbox prevents mounting otherwise and does not supply a mount command with the runtime.