Make file_type to recognize more python files
The `file_type` is just based on file extension, and maybe not by its
actual file type. This will also allow backup files, as in: `*.py.bak,
*.py.orig, etc`, by checking the last 2 ext's.

It will also look for python in the “shebang”. This is done when loading
a file, and when saving a file - and will update accordingly.

- `get_dots`: get the number of dots to use in `ext`
  (ie. foobar.py.orig will also pick up 'py')
- `_has_env_in_shebang()`, returns a bool if env is found
- `has_python_file`:  returns True if `python` is found

Closes #3220